refactor(savedplaces): create <SavedPlaces> component to mark saved places on the map
improve code readability; possibly improve performance
